+++ src/main/kotlin/model/domain/state/BlackTurn.kt
@@ -1,6 +1,6 @@
 package model.domain.state
 
-import model.domain.rule.OmokForbiddenRule
+import model.domain.rule.OmokForbiddenRuleAdapter
 import model.domain.tools.Board
 import model.domain.tools.Location
 import model.domain.tools.Stone
@@ -8,5 +8,5 @@ import model.domain.tools.Stone.BLACK
 
 class BlackTurn(board: Board) : Turn(board) {
     override val stone: Stone = BLACK
-    override fun isForbidden(location: Location) = OmokForbiddenRule(board, stone).isForbidden(location)
+    override fun isForbidden(location: Location) = OmokForbiddenRuleAdapter(board, stone).isForbidden(location)
 }
